refactor: move each KvsEngine implementation to its own module
